+++ src/context/AppContext.tsx
@@ -1,32 +1,4 @@
 import { createContext } from 'react';
-import { OperationRegistry } from '../features/blockchain/services/registry';
-import { WalletStore } from '../features/blockchain/stores/walletStore';
-import { TextStreamStore } from '../core/stores/textStreamStore';
-import { ToolCallStreamStore } from '../core/stores/toolCallStreamStore';
-import { MessageHistoryStore } from '../core/stores/messageHistoryStore';
-import { ModelConfig } from './AppContextProvider';
-
-export type ExecuteOperation = (
-  operationName: string,
-  params: unknown,
-) => Promise<string>;
-
-export type AppContextType = {
-  errorText: string;
-  setErrorText: React.Dispatch<React.SetStateAction<string>>;
-  isReady: boolean;
-  setIsReady: React.Dispatch<React.SetStateAction<boolean>>;
-  isRequesting: boolean;
-  setIsRequesting: React.Dispatch<React.SetStateAction<boolean>>;
-  registry: OperationRegistry<unknown>;
-  modelConfig: ModelConfig;
-  executeOperation: ExecuteOperation;
-  walletStore: WalletStore;
-  messageStore: TextStreamStore;
-  toolCallStreamStore: ToolCallStreamStore;
-  progressStore: TextStreamStore;
-  messageHistoryStore: MessageHistoryStore;
-  isOperationValidated: boolean;
-};
+import { AppContextType } from './types';
 
 export const AppContext = createContext<AppContextType | undefined>(undefined);
